Add function calls, make expressions stand as statements on their own
Function calls are identified by a '$'. Function calls fill the same
parsing role as integers and identifiers. An expression by itself ending
with a ';' are parsed as expressions. Mainly did this to allow function
calls to stand on their own since they're only really parsed as part of
an expression.