Refactor almost all code to lex and parse in different stages, implement beginnings of ast generation
Since information like the name of variables and the value of integers
needed to be stored for ast generation, I had to refactor the way that
lexing took place. Tokens couldn't just be a type of token, they needed
to store information about the token. So the token type was changed to a
struct with some fields that aren't used for all token types. In order
to make the lexer easier to work on separately to the parser, I
separated the two stages, so now the lexer creates a queue of tokens
that the parser then accesses to parse the program. Since I've gotten
some criticism for using global variables, I made the parsing and lexing
functions take the token stream as an argument, rather than making it
available to all functions. This gets a bit annoying but I suppose it
makes the code safer. I also had to go thrpugh the code and free tokens
when they were no longer in use to prevent memory leeks since they are
dynamically allocated. Some ast generation functions were also written.
A parent statemtn struct is created by the parse_P function and the
pointers to relevant child nodes are then passed to the functions called
by the parent function. These child functions then populate the ast. The
code was also split up into multiple source files with one header file
for any shared functions. This should make working on the project a bit
more modular. This is why there is now a Makefile, as this is the
recommended way to compile the program.