Add print_stream function to analyse what is produced by the lexer. Stop freeing 'word' in scan_tokens
The lexer currently isn't working properly due to poor pointer
management. I've been passing pointers to functions and then reassigning
them within the function and expecting the original pointer to change.
This doesn't work because the actual pointer being passed to the
function is passed as a copy of the pointer, its memory address is
different to the original address sroting the pointer. I stopped freeing
'word' because I had been assigning this memory to the 'name' field in
identifier tokens, which would then be read as garbage after freeing it.