Fix parsing bugs
Parsing now works. A token in parse_arguments was being freed and then
read, resulting in a garbage token being putback on the stream. The loop
now reads the next token at the end to fix this. Numerous other bug
fixes took place in this commit but I can't remember them all.