Move source and header files to src/ directory
I did this for the sake of having a cleaner repo. It was getting
confusing and messy to have the source files and makefile, todo file,
etc. in the root directory of the repo. I also updated the clean rule in
the makefile to remove all object files rather than just the specific
object files produced by the compilation. This is to prevent errors
occuring when one of the object files is missing.